move floating panel to the design system
