+++ apps/builder/app/builder/features/project-settings/image-control.tsx
@@ -1,6 +1,6 @@
-import { FloatingPanel } from "~/builder/shared/floating-panel";
-import { ImageManager } from "~/builder/shared/image-manager";
 import type { ReactElement } from "react";
+import { FloatingPanel } from "@webstudio-is/design-system";
+import { ImageManager } from "~/builder/shared/image-manager";
 
 // @todo should be moved to shared as its being reused in another feature
 export const ImageControl = (props: {
